Nova Scotia Basic Personal Amount Fixes #395 Reformed the varaibles part code in a more mathematical way
