+++ changelog_entry.yaml
@@ -1,6 +1,5 @@
-- bump: patch
+- bump: minor
   changes:
     added:
-    - Chage names to proper expressions for title, label, reference and folder.
-    - Make format.
+    - Nova Scotia Basic Personal Amount.
