+++ policyengine_canada/parameters/gov/provinces/ns/tax/income/credits/basic_personal_amount/additional_amount/lower_income_threshold.yaml
@@ -1,11 +1,9 @@
-description: Nova Scotia employs this applicable rate for calculating partial basic personal amounts.
+description: Nova Scotia phases its additional basic personal amount out for filers with income above this threshold.
 values: 
-  2022-01-01: 0.06
+  2022-01-01: 25_000
 metadata:
-  unit: currency-CAD 
-  period: year 
-  rate_unit: /1
-  label: Nova Scotia Basic Personal Amount Applicable Rate
+  unit: currency-CAD
+  label: Nova Scotia basic personal amount additional amount phase out start
   reference: 
   - title: 2022 Nova Scotia Personal Tax Credits Return
     href: https://hr.acadiau.ca/files/sites/hr/Payroll/Pensions%20&%20Benefits/NS_TD1_2022.pdf
