+++ policyengine_canada/parameters/gov/provinces/ns/tax/income/credits/basic_personal_amount/base.yaml
@@ -1,9 +1,9 @@
-description: Nova Scotia limits this higher income threshold under the basic personal amount eligibility.
+description: Nova Scotia provides this base amount under the basic personal amount.
 values: 
-  2022-01-01: 75_000
+  2022-01-01: 8_481
 metadata:
   unit: currency-CAD
-  label: Nova Scotia Basic Personal Amount Higher Income Threshold
+  label: Nova Scotia basic personal amount base
   reference: 
   - title: 2022 Nova Scotia Personal Tax Credits Return
     href: https://hr.acadiau.ca/files/sites/hr/Payroll/Pensions%20&%20Benefits/NS_TD1_2022.pdf
