--- policyengine_canada/parameters/gov/provinces/ns/tax/income/credits/ns_basic_personal_amount/additional_amount.yaml
@@ -1,11 +1,0 @@
-description: Nova Scotia provides this additional amount under the basic personal amount.
-values: 
-  2022-01-01: 3_000
-metadata:
-  unit: currency-CAD
-  label: Nova Scotia Additional Personal Amount
-  reference: 
-  - title: 2022 Nova Scotia Personal Tax Credits Return
-    href: https://hr.acadiau.ca/files/sites/hr/Payroll/Pensions%20&%20Benefits/NS_TD1_2022.pdf
-  - title: Worksheet NS428
-    href: https://www.canada.ca/content/dam/cra-arc/formspubs/pbg/5003-d/5003-d-22e.pdf
